// frontend/src/backend.d.ts
export interface Location {
    country: string;
    city: string;
    state: string;
}
// ...
export interface Ngo {
    id: string;
    verified: boolean;
    focusArea: string;
    contact: ContactInfo;
    name: string;
    registrationNo: string;
    location: Location;
    services: Array<string>;
}
// ...
export interface Vendor {
    id: string;
    verified: boolean;
    contact: ContactInfo;
    name: string;
    category: string;
    products: Array<string>;
    location: Location;
}
export interface LuxuryDebtService {
    id: string;
    serviceType: string;
    applicantName: string;
    approvalStatus: string;
    amountRequested: bigint;
}
export interface ContactInfo {
    email: string;
    website: string;
    address: string;
    phone: string;
}
// ...
export interface Hospital {
    id: string;
    contact: ContactInfo;
    radiologyCriteria: string;
    name: string;
    phisoCourses: Array<string>;
    accreditation: string;
    phisoPractices: Array<string>;
    address: string;
    location: Location;
    services: Array<string>;
}